+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Location module
   * Capture and teleport to locations in the flight simulator
   * Manually add and edit locations
+  * Stores also pitch, bank, heading and initial speed
   * Country list according to [List of Country Codes [Wikipedia]](https://en.wikipedia.org/wiki/List_of_ISO_3166_country_codes)
 
 ### Improvements
@@ -15,8 +16,8 @@
 
 ### Bug Fixes
 - The window is properly resized when in minimal mode and the default non-essential button visibility is changed in the settings
-- Selection in the logbook and formation modules is now properly taking reordering of the table into account
-- The formation module is properly updated when a new logbook is opened or created
+- Selection in the Logbook and Formation modules is now properly taking table reordering into account
+- The Formation module is properly updated when a new logbook is opened or created
 
 ### Under The Hood
 - The modules (Logbook, Formation, Location) are now implemented as plugins